refactor: leave page confirm check
